requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.3.0
langchain-chroma>=0.2.0
chromadb>=1.0.0,<2.0.0
python-dotenv>=1.0.0
